Event Manager: Process mouse button events asynchronously
If we process mouse button events synchronously, they will be handled
prior to any events that are a direct result of that button being
clicked. Normally this is not a problem, but if the result of the
click is to cause the active application to change, we are in danger
of processing the event with the script from the inactive app.