Web: Don't treat DOCUMENT_FRAME as a stop-at role for generating ancestors
When generating ancestors, we typically stop at large containers so as
not to be too chatty. We were treating DOCUMENT_FRAME as such a container,
primarily because it was how web documents were exposed to us (prior to
the creation of DOCUMENT_WEB). Now that we have DOCUMENT_WEB, we should
be able to distinguish ARIA's document role (DOCUMENT_FRAME), from a web
document. And because we cannot control where an author might choose to
use this role, we should not take it as a reason to stop generating
ancestors. Example: The FluentUI react dialog applies the document
role to a div which stands between the buttons and the dialog.

Fixes issue 250