Web: Eliminate chattiness from newly-expanded/shown listbox
When a newly-expanded/shown listbox appears and focus is given to
a child option, we generate ancestors. That presentation includes
the name of the ancestor listbox.

In some ARIA authoring, the name of the listbox is the same as the
previously-focused object (e.g. a button or collapsed combobox).
Repeating this name is unnecessarily chatty. This commit checks for
that condition and returns early for the listbox name generation.

Possibly fixes issue #252.