Web: Support presentation of child position for description list values
NOTE: This functionality requires "speak child position" to be enabled
on the "Speech" page in Orca's preferences dialog (it is not enabled by
default).

No user agent provides this, so we have to do some accessibility tree
walking. It seems not too horrible given that we've already cached
things as part of supporting child position for description list terms.

See issue #248