Web: Don't return early when generating ancestors for table descendants
If a user is tabbing among links inside an ARIA grid, we were not
speaking the headers and positions of column headers. In order to fix
this completely, we need to do two things: 1) Generate ancestors and
2) Adjust what information gets generated for the cell information.
The second might have undesired side effects, so the changes are being
split. This commit only addresses the first.

See issue #253