Create 'ancestor' formatType for presentation of objects
The 'focused' formatType has been serving dual purposes, specifying
what to present when the object is already focused and something
changed and what to present about an ancestor of a newly focused
object. This approach was likely done because the information in
both cases is expected to be minimal, and the set of roles for
which we might want distinct already-focused info and as-ancestor
info small to non-existent.

It turns out that small may be true; non-existent isn't. Consider focus
changing among items (e.g. links) descending from ARIA gridcells. If the
column changes, we want to include some of the new cell/column info. And
that info will be different from what we present when some state
changes in the currently-focused Gtk table cell.

This commit merely adds the capability of an ancestor formatType
because having it makes sense in general. It falls back on the focused
formatType, so no end-user-observable behavior should result from this
commit. A subsequent commit will use this new formatType to solve the
aforementioned table cell problem. And as other such instances become
noticed, we can start decoupling "focused" from "ancestor."

See issue #253.