Store row and column info even when the focused object isn't a cell
The logic to determine whether or not the row and/or column header
has changed relies on the values stored when the location changes.
We were only storing those values when the new location was a cell
(or cell-like thing). If instead the new location is a focusable
descendant of the cell, we are at risk of duplicating header info.

This commit fixes that by looking for the most immediate ancestor
cell and storing its coordinates if found.

See issue #253.