Web: Make cell ancestor presentation optional
In issue #253, Orca's failure to present cell ancestor information
when tabbing among interactive descendants of a cell was reported
as a bug. Fixing that bug has resulted in users complaining that
this presentation can be too chatty and should be configurable.
In order to accomplish that without introducing new settings (and
strings, and UI), this presentation now respects the table-related
options found on Orca's app-specific preferences for web browsers.
Those options had been intended just for structural navigation within
tables, but seem to be applicable here. Because they are enabled by
default, users who find the ancestor information too chatty will
need to opt out by disabling that presentation.