Web: Prevent structural navigation exiting current modal dialog
When using browse mode inside of a modal dialog, it should not be possible
to use structural navigation to exit that modal dialog without first
dismissing it. This commit causes structural navigation objects which are
outside of the current modal dialog to not be included in the list of
available objects or reachable via previous/next object.

Fixes issue #254.