Web: Fix issue causing Orca to remain silent with autofocused inputs
When a text field is autofocused, we might get the caret-moved event
prior to the focus event. In all cases, we update the locusOfFocus.
However we were only announcing the new focus to the user in the case
of entries and spin buttons; not editable combo boxes or password
fields. This commit fixes that.

Fixes issue #259.