Fix presentation of Thunderbird new feed wizard
We automatically present static text content in a newly-shown dialog
because that content is not something which would likely gain focus,
and users should not have to flat review it. The code that looks for
the static text content includes a crude heuristic (minimum word count,
which defaults to 3) that attempts to catch and eliminate labels believed
to be labeling something but lacking the proper relationship. In the case
of the new feed wizard, there's static text consisting of only two words.

This commit sets the minimum word count for Gecko-based apps to 1. If
this results in extra text being presented due to the lack of proper
relationship, that will need to be fixed in the Gecko app.

Fixes issue #211