Refine logic related to inclusion/exclusion of description
Orca attempts to avoid presentation of descriptions which are the
same as the name/label. The logic was that if we had a description
and it wasn't the same text as either the name or the label, present
it; otherwise don't. That makes sense in the case where both the name
and label are to be presented. It also makes sense for web content,
where any "label" is exposed as the accessible name. However, for
widgets outside of web content for which we present the label (if
found) and the name (if there is no label), we can incorrectly filter
out a description due to being equal to the (not presented) name.

This commit refines the logic as follows:
* If we're presenting both label and name, the description must be
  different from both.
* If we're presenting label or name and we have a label, the description
  must be different from the label (but can be the same as the name).
* If we're presenting label or name and only have a name, the description
  must be different from the name.

Fixes issue #218.