Handle exception when using AtspiTableCell
Commit cb105d4d21c added the use of the TableCell interface as the
preferred means to get the table cell's coordinates. It returns the
value only if successful, and checks that the interface is available,
but did not handle exceptions, such as the to-be-debugged one seen when
attempting to call Atspi.TableCell.get_position on the first row in
Caja. This commit adds exception handling so that we fail gracefully.