Handle layout-only descendants in table cells
When attempting to get the displayed text of a table cell that has no
children, we look for the text and fall back on the accessible name.
But if that table cell does have children, we instead would iterate
through them and concatenate their text/names. We were not handling the
case where the table cell does have at least one child, but all its
children are purely for layout (lacking text and names). We should
handle such a table cell as if it had no children. This commit makes
that change.

See related Nautilus/Files issue #2460.