Handle implementations which do not give us all the headers for a cell
This change retrieves all the headers for a cell for implementations
which do the following:

1. Only give us the innermost/closest header for a cell
2. Support returning the header of a header

Firefox is such an implementation; Chromium is not. While Chromium does
give us all the headers for a cell (good; nothing to work around), it
does not support returning the header of a header. Thus when navigating
in a nested row/column header, we cannot present the changed outer
header. That should be fixed in Chromium.

Fixes issue #269 (to the extent we can do so in Orca)