Use presentMessage rather than presentObject in onNameChanged
When we use presentObject, the formatting string usually includes
the label and/or name. But it does not for all roles, such as the
PUSH_BUTTON role. Using presentMessage with the event.any_data (if
present) is consistent with what we do for description changes and
will ensure the name change of the currently focused/active item
is spoken.

See issue #255.