Don't present ancestry of newly-selected page tabs
The ancestry of newly-selected page tabs typically is flagged as layout
only and is thus not presented by Orca. If, however, the ancestor has
a name, it is no longer considered layout only. Firefox's page tabs are
contained in a named toolbar. As a result, when the active browser tab
is changed via keyboard shortcut, we present not only the name of the
now-active/selected tab but also the parent toolbar.

There's probably no good reason to ever present the ancestry of a
newly-selected page tab, so enforce that.

See issue #276.