Web: Treat ARIA's tab role as children-presentational
According to the ARIA spec, the tab role's children are presentational
and therefore user agents should not expose its children. Chromium
follows this; Firefox does not. As a result, we risk double-presenting
the label of a page tab in Firefox when in browse mode. Fix this by
enforcing the children-presentational within Orca.

Fixes issue #288.