web: refresh key grabs after updating focus
When the web script is reactivated, refreshKeyGrabs is called, but
locusOfFocus might not yet be set to the currently active control, so the
script doesn't know that structural navigation is active, meaning that at-spi
isn't instructed to add the key bindings for structural navigation. Later,
onFocusedChanged is called, and it might update locusOfFocus without notifying.
We need to ensure that the AT-SPI key grabs are updated in this case.