Web: Stop treating ROLE_DOCUMENT_FRAME as an indication we're in a web document
Historically browsers used ROLE_DOCUMENT_FRAME for the web document's role.
As a result, we had to support that as well as ROLE_DOCUMENT_WEB when
determining if we're in document content (i.e. and can use browse mode and
structural navigation commands). Now that browsers have been using the
correct role (ROLE_DOCUMENT_WEB) for years, we can stop supporting the old
role.

No longer supporting the old role will also fix bugs in which there is
browser UI which exposes a document-frame role as an ancestor of a
dialog or alert. See issue #293.

In addition to the above change, more debugging output was added which
would have flagged this issue more clearly and which should make it
easier to get to the bottom of bugs like "Orca sometimes uses structural
navigation commands in the location bar."