Web: Fix jumping in sites caused by ARIA selection-changed events
We were treating selection-changed events from ARIA as relevant
unconditionally. This can lead to the presentation of unwanted items
from carousel and might also lead us to update our location to a
carousel item. Don't treat these events as relevant if they come
from an element that we are not on or in.

Note that this change is being conservative due to the upcoming GNOME
release, and only handles specific roles (currently just page tab list)
based on real-world use cases. We can try being more aggressive later.