Fix double-presentation of indentation information
Some synthesizers will verbalize the leading whitespace based on settings
outside of Orca. If the user has enabled this verbalization at the
synthesizer level and also enabled it in Orca, there can be double
presentation of the indentation information. Therefore, if it's enabled
in Orca, strip it from the string to be sent to the synthesizer.