Use the GTK specific scripts for GTK 4 apps as well
This maps the toolkit identifier of GTK to gtk, which allows to use
GTK specific behaviors in GTK 4 apps.
The main motivation for this change is to allow reusing the alert reading
behavior for GTK. It does nothing for GTK 4 just now, but I plan on allowing
applications to send the relevant showing notifications, so, for example,
an appearing AdwToast will be automatically read.