Remove the gnome-screensaver-dialog script
This script was created back in 2008 to handle one specific scenario
in a very-specific way, namely to look at the children of this object:

   event.source.parent.parent.parent

It's now 15 years later. If anyone is still using gnome-screensaver,
and if that dialog still exists, we'll only do the correct thing if
1) event.source.parent.parent.parent exists and 2) is the same
object with the same children. This seems rather unlikely, that said:
If removing this script impacts anyone, they can provide us with
information on what to install and how to reproduce the problem. At
which point we can do a more robust solution to restore the missing
functionality (including documenting what the parent.parent.parent of
the event source is).