Don't speak "frame" role for docked windows
When one switches to the top/bottom panel in mate-panel, we announce the
name, e.g. "Top Panel" followed by the role "frame." While the top panel
reports being a frame and fires window-related events, it functions as a
panel or even a toolbar. Since the "frame" role doesn't provide any useful
information, we should eliminate this chattiness.

Gtk windows expose the window type as an object attribute. The panels have
a type of "dock." We can use this to stop speaking the "frame" role for
mate-panel. Operating under the assumption that all docked windows should
not be announced as typical windows, make the change the default behavior.