Remove gnome-panel script
Note that, unlike other MATE-forked apps, we never made a mapping for
mate-panel to use the gnome-panel support. Thus mate-panel access has
been provided by Orca's default support. In addition, the only thing
the gnome-panel script did was present the name of the top/bottom panel
using the system voice rather than the default voice because the system
voice should be used for additional information that is not displayed.
And the panels' names are not displayed (unlike typical, undocked,
frames). While we could -- and still can -- change the voice used for
the mate-panel panel names, doing so would be a new behavior that no
MATE user has asked for. So just delete the gnome-panel script.