Update LibreOffice script for new spelling dialog accessibility tree
Unfortunately we have to use heuristics to identify the various
components of a spellcheck dialog because there is no ATK/AT-SPI API
specific to spellcheck. Object attributes would be better than nothing,
so https://bugs.documentfoundation.org/show_bug.cgi?id=155447 was filed.

In the meantime, update the LO script in response to changes made in the
LO spellcheck dialog.

Note: There is other new weirdness like panels claiming focus leading
to unwanted chattiness. That will be addressed in a separate commit.