Remove the gnome-window-properties script
The gnome-window-properties script was added many years ago to do exactly
one thing: prevent Orca from being chatty due to all the unrelated labels
in that dialog. Since then, logic has been added to Orca's unrelatedLabels
function to filter those labels out for all dialogs without the need for a
custom script.

This can be seen when opening the mate-window-properties dialog: We are
not mapping mate-window-properties to the gnome-window-properties script,
and Orca is not especially chatty. There is one label Orca is currently
not filtering out, and logic can be added to handle labels like that
across the board should it prove a common occurance. In the meantime, if
https://github.com/mate-desktop/mate-control-center/issues/703 is fixed,
not only will Orca stop speaking that one label when it shouldn't, it
will start speaking it when it should.