Stop using Orca's custom character names for non-math content
Orca sometimes speaks the same character differently. That is because
sometimes it sends the character unmodified to speech-dispatcher and
other times it uses its own custom dictionary. The presentation should
be consistent, and the best (easiest, more performant) way to accomplish
this is to stop modifying the character names.

The end goal is to eliminate chnames.py completely. Before we do that,
we should be sure there are no unexpected side effects. Therefore,
start just by not using the chnames dict for non-math content.

See issue #313.