Add logic to filter out windows which claim to be active when they're not
Windows which can claim to be active when they're not include:
* Descendants of mutter-x11-frames which duplicate existing frames
* The desktop frame of some window environments
* Certain Electron apps (e.g. Slack)

Add logic to eliminate the mutter-x11-frames frames. Tweak the logic
related to eliminating the desktop frame to look for the window-type
object attribute which was added a few years ago. Finally create a
list of "suspect" apps (currently just Slack). This should improve
Orca presenting the correct window when started, as well as identifying
which events can be ignored when ignoring event that are not from the
active app.