Web: Improve presentation of focusable cells which lack accessible names
Focusable objects should have accessible names provided by the author
or calculated by the browser. Chromium-based browsers typically do the
fallback calculation; Firefox does not. We were falling back on the
text content of the cell in this case, but if the cell's content is
inside some other object, such as a div, we were not presenting anything.

We now look for the first descendant which has displayed text and
present that. We deliberately are not doing further repairs (such as
piecing together the contents of the descendants) because that might
lead to unexpected chattiness.

Fixes issue #319.