Web: Eliminate listbox chattiness with ARIA comboboxes
If focus is in an editable combo box and the user types in the input,
the web app should update the active descendant which will in turn
cause the browser to fire a focus event on the newly-selected list item.
We do not want to present the ancestor list box every time the user
types something and the active descendant changes. Therefore, if we get
a focus event for a list item, and the previous focused object was a
combo box with a controller-for relation pointing to the list item's
listbox, do not generate the ancestors.

Fixes issue #317.