Interrupt on focus change
Normally, we interrupt speech when the user presses a key. As a consequence,
we interrupt presentation of the current focus after the focus changes,
because most often, it is done via user initiated keyboard actions.

In GTK 4 apps under Wayland, however, we don't receive the keyboard events,
so among many things, we don't get an opportunity to interrupt our speech.
As a result, Orca is speaking every focus change, even the user is long
ago focused on a different control.

To fix this, interrupt the speech on focus changes ourselves.

Also, make sure that we allow some focus changes to be spoken, for example,
when you focus a particular menu directly, read its name and the menu item
you landed on.

Fixes #327