Keep track of "dead" objects when exception is encountered
AXObject handles exceptions so that we fail gracefully when an object
is being destroyed. But up to this point that was all we were doing.
As a result, we might try to obtain multiple properties (e.g. role,
name, description, etc.) for the same dead object. That's pointless
and not good for performance.

Have AXObject look at the exception and the exception is due to that
object no longer existing, or the app no longer existing, add the
object to a list of known dead objects and return early for any object
in that list.

Note that this currently is only used in AXObject functions. Using it
in the EventManager will be done in a subsequent commit.