Also sanity check the active window in default script
onActiveChanged and onWindowActivated have logic based on whether
or not the object firing the event can be the active window. We
added logic a while ago to eliminate mutter frames as the active
window, so we need to do the sanity checking to locate the real
frame and use the result in the subsequent logic.