Use the new Atspi clear_cache_single.
Normally we do not want to clear the Atspi cache for an entire subtree;
just a single object (e.g. to refresh its state set). Prior to Atspi
v 2.49.1, we only had clear_cache which is recursive. Given that we
periodically need to clear the cache of a window, that can have
serious performance implications. Now that clear_cache_single exists,
try it first. If we get an exception, fall back on the old clear_cache.