Don't cache roles
This is another speculative fix for Orca saying weird things related
to object roles. It might also fix the occasional "Cannot read this
email" reports. The theory behind this change is that objects are
starting out with a default role and that role is then being changed
by the app or toolkit. Should this occur, we'd have bogus information.

Having bogus information could not only result in Orca speaking the
wrong role, but also conclude that a document is not a document (which
would break caret navigation).