Fix Orca not speaking the newly-selected item when deleting message
Orca was not speaking the newly-selected item when deleting the message
because the is-same-object heuristic passed in `locusOfFocusChanged`.

In the case of deleting the selected item in a message list, there's a
good chance that the heuristic will pass because you'll likely land on
a cell that occupies the same space, has the same role, and might even
have the same name, path, etc., as a consequence of the current one
no longer existing. Address this by not applying the heuristic in the
case of table cells, checking true equality instead. That got Orca
talking, but...

Evolution is firing `active-descendant-changed` events for ... who knows
what? Cells with names like "20/00/-499498928 13:28". Having Orca say
that is arguably worse than remaining silent. Happily, the existing
AXObject debugging checks flagged this condition elsewhere. So fix the
bogus descendant by moving the relevant check to a new function,
`get_active_descendant_checked`. Have our existing checker call it, and
only for containers which claim to manage their descendants. (Nearly
all of the errors I've been seeing fall into that category anyway.)
To fix the presentation, use this function in Evolution's script.
If it detects bogusness and can get a valid cell instead, present
the valid cell. That got Orca saying something that actually makes
sense, but...

Orca would start speaking the new name of the frame (e.g. "Inbox 152
unread") but get cut off before it was finished and start speaking
the newly-selected item. Having part of the frame presentation
immediately followed by the newly selected item can be hard to
mentally process. Either we don't speak the frame name in this case,
or we let it finish. I personally think the unread count is useful.
If others disagree, we can change it in a later commit.

The reason the frame presentation was getting cut off is due to the new
need to try to cut off speech when we get no input events (think Gtk 4
+ Wayland). Fix this by adding an optional `event` argument to the newish
`shouldInterruptForLocusOfFocusChange`. For `active-descendant-changed`,
defer to the active script's `stopSpeechOnActiveDescendantChanged`
function. Evolution's already returns false. We may need to tweak that
in a Gtk 4 + Wayland + Evolution world, but that too can be done in a
later commit.