Fix regression in Firefox live region presentation
Because Thunderbird is spamming us horribly when one expands or
collapses a thread of messages, we started filtering out the
associated events. It turns out that sometimes, those events are
the very same ones we get from Gecko for live regions. Therefore
be sure the events are from Thunderbird before we filter them out.