Web: Attempt to make _canHaveCaretContext more performant
We didn't use to cache results because a number of the checks done
can change. For instance something empty or "useless" might become
non-empty and non-"useless" in an interactive app. But this lack
of caching seems to be costing us in performance. Therefore:

1. Start caching the decisions, BUT
2. Clear the cache immediately before, and at the end of:
   * Getting the char, word, sentence, line, object contents
   * Finding the first caret context
3. Add some performant checks to the top (e.g. editable, focusable,
   and landmarks can always have the caret context).
4. Re-order the remaining checks based on likelyhood and performance
   cost (e.g. checking for aria-hidden is less performance due to
   our having to use a non-cached AtspiObject attribute)

Hopefully this will shave at least some time off of the lagginess
reported in issue #355.