Event manager: Implement support to suspend certain accessibility events
This change was inspired by gnome-shell's event floods, but arguably
applies to other cases.

If we get an event related to window creation, activation, deactivation,
or destruction, there is a reasonable chance we'll also get a bunch of
events for the creation/removal of objects in that window along with
possible state changes (showing, sensitive, etc.) in preparation for
making that window available/unavailable to the user. If the app's
accessibility tree is large enough, that can result in a flood.

We've been dealing with these floods by immediately rejecting the event
in question. That has helped, but even that processing takes time and
can make Orca laggy. Therefore:

* Create a means to temporarily suspend/unsuspend those events by
  deregistering/reregistering them with the Atspi registry.
* Use this new functionality for window-related events to stop
  listening for children-changed events. Future events will be
  done incrementally.

This change makes using gnome-shell's window switcher signficantly
more performant.