Remove obsolete rhythmbox script
The rhythmbox script served exactly one purpose: to present a rating
widget which had the "unknown" role. We typically ignore objects with
the "unknown" role so the likelihood of this script actually doing
anything is low. And if it was doing something, then the real fix is
to give the rating widget a proper role. Many roles would be "good
enough," but there's a dedicated "rating" role in ATK/AT-SPI2.

https://gitlab.gnome.org/GNOME/rhythmbox/-/issues/2052