Remove obsolete gnome-search-tool script
The main purpose of the gnome-search-tool script was to prevent
Orca and the desktop from completely locking up due to event floods.
It did not make things usable, and the need to unregister from much-
needed accessibility events made other apps less accessible. In
addition, gnome-search-tool no longer exists in GNOME. It persists
in MATE as mate-search-tool, but the problems from the GNOME 2 days
still exist and can be reproduced with other accessibility clients
without Orca running. Those problems should be addressed either in
mate-search-tool, and/or Gtk, and/or Atspi.

https://github.com/mate-desktop/mate-utils/issues/352
https://gitlab.gnome.org/GNOME/at-spi2-core/-/issues/133