Web: Attempt to improve performance retrieving clickables
This takes advantage of the new collection's utilities to speed
things up a bit. Much of the lag was coming from the application
of non-performant predicate functions done after retrieving a
ton of results. By using the collection interface for more of
that logic, the predicates we still need to apply should run
faster.