Flat review: Prefer dialog over top level object for context window
This is a speculative fix for a reported regression that Orca was
no longer properly performing flat review in VS Code dialogs. The
regression was bisected to 284cd4f9 which ensured we were not
basing our decision on what was the top-level window solely on
roles. Qt, for instance, uses the frame role for a lot of things
which are not frames (in the ATK/AT-SPI2 sense). Because the flat
review logic uses topLevelObject to find the containing window,
any app whose dialogs descend from the top level (as opposed to
being their own top-level object) could have weird flat-review
presentation as a result of 284cd4f9.

In order to (hopefully) fix that, change the flat review logic to
use frameAndDialog instead. The frame should be the top-level object.
If there is a dialog that is not a top-level object, it will also
be returned. Then prefer the dialog over the frame when building
the flat review context.