Quadruple check we have an app when Firefox destroys and re-creates itself
We already have to triple check the accessible application associated
with Firefox because, for reasons yet to be determined, it seems to
like to destroy and re-create itself. For the most part, we seem to be
handling this. But there appears to be a timing issue where we don't
yet have the updated accessible application, though we have a new
script for it. Handle this condition and add some more debugging to
catch other oddness.