Don't attempt to pan via keyboard command without braille enabled
Orca has an unbound command which makes it possible to pan a braille
display using a traditional keyboard rather than the braille display
inputs. If this command is bound and used, braille support should be
enabled, otherwise there's nothing to pan and an unhandled exception
is raised. Fix that by checking for braille support if the input
event is a KeyboardEvent (as opposed to the BrailleEvent we'd expect).