Don't interrupt speech for focus change for editable cell
Thanks to the fact that we can not count on input event notification
(e.g. Gtk4 + Wayland), we now interrupt speech in response to focus
changes, making exceptions for specific cases. One case we need to
handle is when the user is on an editable table cell and presses
return to edit that cells contents. Any information about that cell
gets cut off in response to the new text object claiming focus.

Handling this should ensure Orca completes the prompt related to
rebinding key commands in its preferences dialog.