Speculative fix for MATE's window switcher not being announced
Commit 0a7ac51472300c9199dd756ba7268c19a982aba6 suspended more events
in response to window activation/deactivation as well as when the
busy state changes. That commit appears to have regressed the
presentation of MATE's window switcher.

This commit removes suspension of both showing and text insertion
events on the suspicion that one of those is what we use to present
the newly-selected item. Once we determine which is the needed
event, we can resume suspending the other to regain some performance.