Added object navigation functionality.
This will allow the user to navigate through a hierarchy of UI
objects, mostly based on the object hierarchy as it appears to Orca.

Added the following commands:
* Commands for moving to the parent, first child, next and previous
  sibling. Bound to CTRL+Orca+arrows.
* A command for acting on the current object with navigator focus,
  usually performs a click. Bound to CTRL+Orca+Return.
* A command for toggling the simplification of the object hierarchy.
  Bound to CTRL+Orca+S.

Note that the original change is being landed with a number of
updates to bring the object navigator in line with other changes
currently taking place. If these changes introduced regressions in
the original functionality, blame your humble maintainer; not the
author of this feature.