Flat review: Update the current context to a new object of interest
The FlatReviewPresenter and ObjectNavigator try to remain in sync
with each other. But there was one scenario not yet handled:

1. User enters flat review and reviews some on-screen objects
2. User switches to the object navigator and explores some more
3. User performs a flat-review command

In this particular scenario, the flat-review context was never
destroyed. And destroying it when modes change can be non-performant.
This commit attempts to address that by updating the current context
to the new object of interest.

Accomplishing this can be tricky because the FlatReviewPresnter's
objects and the ObjectNavigator's objects can be dramatically
different. The FlatReviewPresenter objects are typically text-based.
Thus if a tree item  has descendant text objects which contain that
item's displayed text, the text objects will be in the flat-review
context whereas the tree item itself will not. Another difference
is due to the fact that the ObjectNavigator lets the user explore
all objects -- including those off-screen, such as items in a closed
menu. Those menu items will not have any representation in the
flat-review context because they are off-screen, and flat-review by
definition is a way to review on-screen objects.

In order to do the best we can with this situation, do the following:

1. If the object of interest from the object navigator is in the
   flat review context, update the flat review context to the
   navigator object.
2. If the object of interest from the object navigator is not in
   the flat review context, but is an ancestor of something in
   the flat review context, update the flat review context to
   that descendant.
3. Otherwise don't update the flat review context.