Attempt to avoid event flood when preferences dialog is destroyed
When the preferences dialog is destroyed, we get an event flood which
appears to be connected to the combo boxes. Even though we ignore those
events, they are apparently enough to make Orca unresponsive for a few
seconds. And most of these events seem to be fired before the previous
window is activated, thus Orca's existing event suspension is not
kicking in until after the flood. Therefore, when the window is being
closed, we now do the following:

* deregister from all the floody events
* do our cleanup
* set a delayed task (currently one second) to re-register the events
* allow destruction to complete

This appears to make the flood go away. Also the window events which
follow the dialog destruction trigger Orca's floody-event suspension.
We just need to stop the flood between destruction and window activation.