Speculative fix for the "screen reader on" message getting cut off
Don't interrupt speech if the locus of focus change has no event.
When Orca starts, it looks for the active window and then the
focused object in that window. In applications where AtspiCollection
is supported, we can find the focused object (if there's one to be
found) a lot faster. Maybe that's why speech is getting cut off?
When Orca sets focus at start up, there is no event, so try to
catch this scenario and not interrupt speech.