Only suspend events from window role if window is from clutter
This is a speculative fix to hopefully avoid a problem caused by
the combination of:

1. MATE failing to emit expected events related to windows
2. The new keyhandling possibly(??) triggering extra window events

The main reason to suspend events temporarily due to an event from
an object with the "window" role (as opposed to the "frame" role),
was gnome-shell event floods. But we're seeing "window" events when
an app creates a popup with a text field (e.g. pluma's goto line
command). Therfore if the object has the "window" role, only suspend
events if the toolkit name is "clutter."