Don't let brlapi crash us
There is something wrong with brlapi in (at least) Fedora 39. To
reproduce the problem without Orca running, try the following in
a terminal:

    python3 -c "import brlapi; brlapi.Connection()"

You may get an error, e.g. with a traceback with messages like
"Connection refused", "Is BRLTTY really running?", something related
to keys being missing, etc. All those are fine (or at least not
severe). If, however, you see this:

    double free or corruption (!prev)
    Aborted (core dumped)

That's a severe error in brlapi, which is currently crashing Orca.

This commit attempts to catch this specific brlapi error and not
crash. It cannot make braille support magically work, but it's
better than nothing.

See issue #386 for details.