Don't interrupt speech in response to window:deactivate events
The new logic related to cutting off speech in response to focus
changes hopefully renders doing this work in onWindowDeactivated
unnecessary. In addition, this interruption will cut off the
presentation of "command not found" in gnome-shell's run dialog due
to https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/6925.

Note that this change is insufficient to actually presenting the
label.

See issue #388.