Don't suspend name-change events
Suspending name-change events in response to window (de)activation can
help with certain event floods, but can also cause us to not present (at
least) the "command not found" label in gnome-shell's run dialog.

The event suspension was being triggered unexpectedly due to the events
we receive when Return is pressed but the dialog not closed. For details
see https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/6925.

This change, on top of the previous three commits, should cause Orca
to present the "command not found" label.

See issue #388.